% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_table_consistency_checks.R
\name{assert_join_covered}
\alias{assert_join_covered}
\title{Assert LHS of join is covered by RHS}
\usage{
assert_join_covered(df1, df2, join_col)
}
\arguments{
\item{df1}{a data.frame (local or remote).}

\item{df2}{a data.frame (local or remote - must be consistent with \code{df1}).}

\item{join_col}{the column(s) to join by. Uses the same syntax as dplyr, so
you can use a named vector to join tables with different column names.
For example: \code{join_col = c(col_in_df1 = "col_in_df2")}}
}
\value{
Errors if assert fails. Else, number of rows invisibly.
}
\description{
When joining two tables, we may want to be sure that every record
on the left table will be joined to a record on the right table.
}
\details{
Asserts every value in join_col(s) of \code{df1} is found in \code{df2}.
If assert passes (without error) then no records in \code{df1} will be lost during
an inner join to \code{df2}.
}
