% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_table_consistency_checks.R
\name{assert_size_comparison}
\alias{assert_size_comparison}
\title{Assert comparison of table sizes}
\usage{
assert_size_comparison(df1, direction, df2)
}
\arguments{
\item{df1}{a data.frame (local or remote) for left-side of comparison.}

\item{direction}{mathematical operator for comparison.}

\item{df2}{a data.frame (local or remote) for right-side of comparison.}
}
\value{
Errors if assert fails. Returns number of rows invisibly if passed.
}
\description{
Asserts direction applies to number of rows in both tables.
}
\examples{
data(iris)
assert_size_comparison(head(iris), "<", iris)

}
