% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_consistency_checks.R
\name{assert_size_missing}
\alias{assert_size_missing}
\title{Assert number of missing values}
\usage{
assert_size_missing(df, col_name, direction, size)
}
\arguments{
\item{df}{a data.frame (local or remote).}

\item{col_name}{the name(s) of columns to count distinct values/combinations
of. Errors if any \code{col_name} is not a column name of \code{df}.}

\item{direction}{mathematical operator for comparison.}

\item{size}{numeric value to compare against size.}
}
\value{
Errors if assert fails. Else, number of missing values invisibly.
}
\description{
Assert number of missing values
}
\examples{
data(iris)
assert_size_missing(iris, "Petal.Length", "==", 0)

}
