% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_consistency_checks.R
\name{check_all_unique}
\alias{check_all_unique}
\title{Check every entry is distinct}
\usage{
check_all_unique(df, col_name)
}
\arguments{
\item{df}{a data.frame (local or remote).}

\item{col_name}{the name(s) of columns to count distinct values/combinations
of. Errors if any \code{col_name} is not a column name of \code{df}.}
}
\value{
T/F whether every entry in col_name (or the combination of col_names)
is distinct.
}
\description{
Check whether every value is a column (of across a set of columns) is
distinct.
}
\details{
Useful for checking whether a join will be one-to-one. If both tables have
every entry distinct on the columns used for the join, then the join will
be one-to-one.
}
\examples{
# example code

}
