% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_table_consistency_checks.R
\name{check_join_covered}
\alias{check_join_covered}
\title{Check LHS of join is covered by RHS}
\usage{
check_join_covered(df1, df2, join_col)
}
\arguments{
\item{df1}{a data.frame (local or remote).}

\item{df2}{a data.frame (local or remote - must be consistent with \code{df1}).}

\item{join_col}{the column(s) to join by. Uses the same syntax as dplyr, so
you can use a named vector to join tables with different column names.
For example: \code{join_col = c(col_in_df1 = "col_in_df2")}}
}
\value{
T/F whether every value in join_col(s) of df1 is found in df2.
}
\description{
When joining two tables, we may want to explicitly test that every record
on the left table will be joined to a record on the right table.
}
\details{
Returns T/F whether every value in join_col(s) of df1 is found in df2.
If returns TRUE, then no records will be lost during an inner join.
}
\examples{
data(iris)
check_join_covered(iris, iris, join_col = c(Sepal.Width = "Sepal.Length"))

}
