% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_table_consistency_checks.R
\name{check_no_date_overlap}
\alias{check_no_date_overlap}
\title{Returns T/F for whether any overlap in date columns within groups defined by other columns.}
\usage{
check_no_date_overlap(df, start_date, end_date, group_by_cols)
}
\arguments{
\item{df}{a data.frame (local or remote).}

\item{start_date}{the column containing start dates}

\item{end_date}{the column containing end dates}

\item{group_by_cols}{columns to group-by / search within.}
}
\value{
T/F
}
\description{
If two individuals have date periods that overlap with each other but neither individual
as overlapping date periods when considered on their own
Then group_by_cols = person_id will return TRUE (no overlap when grouped)
but group_by_cols = c() will return FALSE (overlap exists without grouping)
}
