% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_table_consistency_checks.R
\name{check_num_distinct}
\alias{check_num_distinct}
\title{Check number of distinct values}
\usage{
check_num_distinct(df, col_name, direction, size)
}
\arguments{
\item{df}{a data.frame (local or remote).}

\item{col_name}{the name(s) of columns to count distinct values/combinations
of. Errors if any \code{col_name} is not a column name of \code{df}.}

\item{direction}{mathematical operator for comparison.}

\item{size}{numeric value to compare against size.}
}
\value{
T/F whether the number of distinct values in the specified column(s)
is consistent with the comparison.
}
\description{
Returns T/F for whether number of distinct values in col_name
(or a combination of col_names) \code{df} is \code{direction} relative to \code{size}.
}
\examples{
data(iris)
check_num_distinct(iris, "Species", "!=", 1)
check_num_distinct(iris, c("Sepal.Length", "Sepal.Width"), ">", 20)

}
