% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_table_consistency_checks.R
\name{check_size}
\alias{check_size}
\title{Check table size}
\usage{
check_size(df, direction, size)
}
\arguments{
\item{df}{a data.frame (local or remote)}

\item{direction}{mathematical operator for comparison.}

\item{size}{numeric value to compare against size.}
}
\value{
T/F for whether the rows of \code{df} are \code{direction} of \code{size}.
}
\description{
Returns whether number of rows in table is in direction relative to size.
}
\examples{
data(iris)
check_size(iris, "==", 10)
# FALSE - iris has 150 rows


}
