% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_consistency_checks.R
\name{check_size_comparison}
\alias{check_size_comparison}
\title{Check comparison of table sizes}
\usage{
check_size_comparison(df1, direction, df2)
}
\arguments{
\item{df1}{a data.frame (local or remote) for left-side of comparison.}

\item{direction}{mathematical operator for comparison.}

\item{df2}{a data.frame (local or remote) for right-side of comparison.}
}
\value{
T/F whether the size of the two tables obeys the direction.
}
\description{
Returns T/F for whether direction applies to number of rows in both tables.
}
\examples{
data(iris)
check_size_comparison(head(iris, 10), ">", head(iris, 5))

}
