% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_table_consistency_checks.R
\name{evaluation_comparison}
\alias{evaluation_comparison}
\title{Evaluate numeric comparison}
\usage{
evaluation_comparison(val1, direction, val2)
}
\arguments{
\item{val1}{numeric value for left-side of comparison.}

\item{direction}{mathematical operator for comparison.}

\item{val2}{numeric value for right-side of comparison.}
}
\value{
T/F whether the comparison holds true.
}
\description{
Provides evaluation of comparison when operator is given as text.
E.g. evaluation_direction(14, "<", 10) returns 14 < 10 which is FALSE.
}
