% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_dataset.R
\name{explore_report}
\alias{explore_report}
\title{Explore dataset with report}
\usage{
explore_report(
  df,
  id_column = NA,
  target = NA,
  output_file = "df explored",
  output_dir = getwd()
)
}
\arguments{
\item{df}{a data.frame (local or remote)}

\item{id_column}{(optional) the column of a numeric id to be used for
filtering (as per \code{filter_to_limited_number_of_rows}).}

\item{target}{(optional) a column to split the exploration by as per
\code{explore::report}.}

\item{output_file}{(optional) name for the report file - will have the
date-time as a suffix. File extension .html is added automatically.}

\item{output_dir}{(optional) the folder to save the report. Defaults to
current working directory. Errors if you do not have write permission to
this location.}
}
\value{
the file name of the report just created.
}
\description{
Provide a wrapper for the explore::report to simplify use - especially for
large tables and remote datasets.
}
