% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_consistency_checks.R
\name{num_missing_entries}
\alias{num_missing_entries}
\title{Number of missing values}
\usage{
num_missing_entries(df, col_name)
}
\arguments{
\item{df}{a data.frame (local or remote)}

\item{col_name}{the name(s) of columns to check for missing values.
Errors if any \code{col_name} is not a column name of \code{df}.}
}
\value{
the total number of missing values in the columns of \code{df} given by
\code{col_names}.
}
\description{
Counts the number of missing values in a column. If given multiple columns,
sums the missing values across columns.
}
\details{
Designed to work with remote and local tables.
}
\examples{
data(iris)
num_missing_entries(iris, "Species")
num_missing_entries(iris, c("Sepal.Length", "Sepal.Width"))

}
