% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_consistency_checks.R
\name{num_row}
\alias{num_row}
\title{Number of rows (remote)}
\usage{
num_row(df)
}
\arguments{
\item{df}{a data.frame (local or remote).}
}
\value{
The number of rows in the data.frame. Equivalent to \code{nrow} for
data.frames in local R memory. Also works for tables in remote databases.
}
\description{
Provides the same functionality as \code{nrow} but also works on remote tables.
}
\examples{
data(iris)
num_row(iris)

}
