% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_consistency_checks.R
\name{num_unique_entries}
\alias{num_unique_entries}
\title{Number of distinct values}
\usage{
num_unique_entries(df, col_name)
}
\arguments{
\item{df}{a data.frame (local or remote)}

\item{col_name}{the name(s) of columns to count distinct values/combinations
of. Errors if any \code{col_name} is not a column name of \code{df}.}
}
\value{
The number of distinct values in the columns of \code{df} given by
\code{col_names}.
}
\description{
Counts the number of unique values in a column (or set of columns).
Designed to work with remote and local tables.
}
\examples{
data(iris)
num_unique_entries(iris, "Species")

}
